requirements.txt
ultralytics
streamlit
pillow
numpy
fastapi
uvicorn
python-multipart

